# Server/ServerApp/Swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Warehouse",
    "version": "v1"
  },
  "paths": {
    "/api/v1/items": {
      "get": {
        "tags": [ "ItemsApi" ],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 25 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Item" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [ "ItemsApi" ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Item" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Item" }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{id}": {
      "get": {
        "tags": [ "ItemsApi" ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Item" }
              }
            }
          }
        }
      },
      "put": {
        "tags": [ "ItemsApi" ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Item" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Item" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [ "ItemsApi" ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "No Content" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Item": {
        "required": [ "name", "price", "quantity" ],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "name": { "minLength": 1, "type": "string" },
          "description": { "type": "string", "nullable": true },
          "quantity": { "maximum": 2147483647, "minimum": 0, "type": "integer", "format": "int32" },
          "price": { "minimum": 0.01, "type": "number", "format": "double" },
          "discount": { "maximum": 100, "minimum": 0, "type": "number", "format": "double" },
          "finalPrice": { "type": "number", "format": "double", "readOnly": true }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Введіть JWT токен у форматі: Bearer {your_token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}